Added BarTableViewCell as a CustomCell
Added a new custom cell for the tableview to properly display
information about a bar. Previously the image was smaller and
no infrastructure set up to allow for more information to be
displayed in the cell. Currently only showing name, phone number,
and image. But now can easily add more UI components to show
more information.