Added BarType Icon support and changed UI Layout for TableViewCell
Switched the layout of TableViewCell to emphasize the images of
the tableview cells. Felt that was a more important aspect of quickly catching
a glimpse into what the bar is like. Add BarType Icon support with the
aim of implementing functionality to later show an icon based on the
type of bar (bar, lounge, cocktail bar, etc.)