Support for Clicking Cells and Detail Bar Information
Added functionality when a cell is clicked open up a new page showing details about the bar.
Now when a cell is clicked a new view controller opens showing the bar information that was
within the cell in a view controller. Idea is to eventually show more information about the bar
on this screen.