CollectionView Showing Bars added to Bottom of MapView
Continued UI work on the MapView portion of the app. Now displays a horizontal
collection view showing the bar information so the user can navigate the map
while seeing a some details about the bars as well. BarCollectionViewCell nib
along with Cocoa Touch Class for BarCollectionViewCell created for this purpose