Moved network calls off the main thread, optimizations around showing MKAnnotations, fixed bug where on click certain annotations wouldn't show any details
The YelpApi calls are now off the main thread. MKAnnotations not within the maps frame
are not shown to the user. Previously all bars were added to the bars array and then parsed
through to decide which should show on the map. This caused some inconsistencies between
what annotations were shown versus what was in the array and resulted in a bug not showing
information on certain bars. Logic flow changed so that a bar is only added to the bars array
if its within the map view frame