Functionality to animate bar collection view cells on annotation click
Previously the collection view cells would grow as the user scrolled horizontally through the collection
such that the center would imply "focus" on a cell. However, the cells weren't growing when a user
would pick an annotation view. Now when an annotation view is clicked the collection view shows
with the corresponding bar/cell enlarged