Implementation of MVVM (working progress), Dark Mode support, wrapping up draw mode
Started making the move to a MVVM architecture. Created the view model and added a lot of the
heavy lifting into that class. Dark Mode now the default and support provided for making sure colors
change properly.

MVVM is a work in progress. Still learning where certain code should be so will continue to make improvements
on this as I continue to learn